# cmd/api/docs/swagger.yaml
definitions:
...
  domain.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      err:
        type: string
      hint:
        type: string
      status:
        type: string
    type: object
...
  models.ListCart:
    properties:
      bookId:
        type: integer
      cartId:
        type: integer
      price:
        type: number
      quantity:
        type: integer
      userId:
        type: integer
    type: object
...
  models.Newcategory:
    properties:
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /admin/addcategory:
    post:
      consumes:
      - application/json
      description: Add a new category with the provided name , no duplicate allowed
      parameters:
      - description: New category information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Newcategory'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added new category
          schema:
            type: string
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: InternalServerError
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add a new category
      tags:
      - categories
...
  /admin/categorylist:
    get:
      consumes:
      - application/json
      description: Get a list of categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Lists all the categories
      tags:
      - categories
...
  /admin/deletecategory/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing category with the provided ID
      parameters:
      - description: Category ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted category
          schema:
            type: string
        "400":
          description: Cannot parse category_id
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Delete failed" or "Please try again
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a category
      tags:
      - categories
...
  /admin/updatecategory/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing category of provided ID with new category
      parameters:
      - description: Category ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: New category information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Newcategory'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated category
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a category
      tags:
      - categories
...
  /user/addtocart:
    post:
      description: Add a  product cart , if already exist qty++
      parameters:
      - description: Product id
        in: formData
        name: bookid
        required: true
        type: integer
      - description: user id
        in: formData
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add productto cart
      tags:
      - cart
...
  /user/cart:
    get:
      description: Get a user's cart by User ID.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of items in the cart
          schema:
            items:
              $ref: '#/definitions/models.ListCart'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get user's cart
      tags:
      - cart
...
  /user/deleteitem:
    delete:
      description: Delete a product from the cart for a specific user.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      - description: Product ID
        in: query
        name: bookid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Item deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a product from the cart
      tags:
      - cart
...
  /user/updatecart:
    put:
      description: Update the quantity of a product in the cart for a specific user.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      - description: Product ID
        in: query
        name: bookid
        required: true
        type: integer
      - description: New quantity
        in: query
        name: quantity
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Quantity updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update quantity of a product in the cart
      tags:
      - cart
...
swagger: "2.0"